requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
Telethon==1.36.0
pandas==2.2.2
python-dotenv==1.0.1
